requirements.txt
django==1.11.1
djangorestframework==3.6.3
markdown==2.6.8
django-filter==1.0.2
gunicorn==0.14.6
psycopg2==2.7.1
whitenoise==3.3.0
httplib2==0.10.3